[inference] Extract a shared LayerwiseReloadWorkerMixin (no behavior change)
`NewInferenceWorkerWrap` carried the vLLM layerwise-reload lifecycle
(`start_weight_update`/`finish_weight_update`, which initialize/finalize the
reload) and the `conv_weights` SKIP_TENSORS guard (Mamba/Gated-DeltaNet conv1d).
Move both verbatim into a shared `LayerwiseReloadWorkerMixin` that
`NewInferenceWorkerWrap` now inherits.

Pure refactor: the new inference path behaves identically. The mixin exists so
the legacy `WorkerWrap` path can reuse the same lifecycle next.
